// src/components/ui/practice/itr/itr-7/components/itr-7-schedule-80ia.types.ts
export interface ITR7Schedule80IAData {
    schedule80IA: {
        section_80ia_4_i: { // a
            undertaking1: number;
            undertaking2: number;
        };
        section_80ia_4_iv: { // b
            undertaking1: number;
            undertaking2: number;
        };
        section_80ia_4_v: { // c
            undertaking1: number;
            undertaking2: number;
        };
        total_deduction: number; // d
    };
}